requirements.txt
# Web Framework
Flask==3.0.2
PyYAML==6.0.1
Flask-WTF>=1.2.1
wtforms==3.0.1
bootstrap-flask>=2.0.0

# Docker Integration (to monitor logs)
docker>=7.0.0
apscheduler>=3.10.4

# Networking & Notifications
requests==2.31.0
python-dotenv==1.0.1

# Cryptography (Standard library hashlib is usually enough, 
# but these are helpful for advanced anchoring)
cryptography==42.0.5

# Development / Utility
python-dotenv==1.0.1
